// src/stats/types/team-game-logs.ts
export interface TeamGameLogsParams {
	dateFrom?: string;
	dateTo?: string;
	gameSegment?: string;
	lastNGames?: number;
	leagueID?: string;
	location?: string;
	measureType?: string;
	month?: number;
	oppTeamID?: number;
	outcome?: string;
	pORound?: number;
	perMode?: string;
	period?: number;
	playerID?: number;
	season?: string;
	seasonSegment?: string;
	seasonType?: string;
	shotClockRange?: number;
	teamID?: number;
	vsConference?: string;
	vsDivision?: string;
}

export interface TeamGameLogsRow {
	seasonYear: string;
	teamId: number;
	teamAbbreviation: string;
	teamName: string;
	gameId: string;
	gameDate: string;
	matchup: string;
	wl: unknown;
	min: number;
	fgm: number;
	fga: number;
	fgPct: number | null;
	fg3m: number;
	fg3a: number;
	fg3Pct: number | null;
	ftm: number;
	fta: number;
	ftPct: number | null;
	oreb: number;
	dreb: number;
	reb: number;
	ast: number;
	tov: number;
	stl: number;
	blk: number;
	blka: number;
	pf: number;
	pfd: number;
	pts: number;
	plusMinus: number;
	gpRank: number;
	wRank: number;
	lRank: number;
	wPctRank: number;
	minRank: number;
	fgmRank: number;
	fgaRank: number;
	fgPctRank: number;
	fg3mRank: number;
	fg3aRank: number;
	fg3PctRank: number;
	ftmRank: number;
	ftaRank: number;
	ftPctRank: number;
	orebRank: number;
	drebRank: number;
	rebRank: number;
	astRank: number;
	tovRank: number;
	stlRank: number;
	blkRank: number;
	blkaRank: number;
	pfRank: number;
	pfdRank: number;
	ptsRank: number;
	plusMinusRank: number;
}

export interface TeamGameLogsResponse {
	teamGameLogs: TeamGameLogsRow[];
}